chore(subgraphs/flt): refactor
