+++ subgraphs/flt/tests/.latest.json
@@ -1,4 +1,4 @@
 {
     "version": "0.5.1",
-    "timestamp": 1657263523312
+    "timestamp": 1657506354897
 }
